+++ elems/Image.lua
@@ -72,7 +72,7 @@ function Image:Draw(x, y)
     elementSize.paddingH, 0, "data/ui_gfx/decorations/9piece0_gray.png")
     local clicked, right_clicked, hovered = GuiGetPreviousWidgetInfo(self.gui.guiobj)
     if hovered and self._config.onHover then
-        self._config.onHover(self)
+        self._config.onHover(self, self.gui.state)
     end
     GuiZSetForNextWidget(self.gui.guiobj, self.z)
     if self._config.colour then