+++ elems/VLayout.lua
@@ -58,7 +58,7 @@ function VLayout:Draw(x, y)
     end
     if hovered then
         if self._config.onHover then
-            self._config.onHover(self)
+            self._config.onHover(self, self.gui.state)
         end
         self.useHoverConfigForNextFrame = true
     else self.useHoverConfigForNextFrame = false end