+++ CHANGELOG.md
@@ -1,1 +1,16 @@
 # Changelog
+
+## Initial release 1.0.0 - 2022-10-8
+### Added
+- Button element.
+- Checkbox element.
+- Layout elements.
+- Image element.
+- ImageButton element.
+- Progressbar element.
+- Slider element.
+- Text element.
+- TextInput element.
+- Gui class.
+- GuiElement class (only used internally).
+- Images and assets used for gui elements.